Major rework
moved back to Berkeley DB Base API while changing the model to something inspired by RDF. I liked the DPL but it apparently doesn't allow byte[] fields in Entity object and I thought that wrapping a String around these was an ugly hack. The object in RDF statements is only a string so there's no chance of making implications and so on... but eventually it'll happen.

Should add some documentation before I forget all I've done :)